% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_attribute.R
\name{get_edge_attribute}
\alias{get_edge_attribute}
\title{Get the Attribute Values of Selected Edges}
\usage{
get_edge_attribute(
  semPaths_plot,
  edges = NULL,
  attribute_name = NULL,
  check_direction = TRUE
)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions.}

\item{edges}{A character vector of
the edges. Each edge is denoted by
\code{lavaan} model syntax, such as \code{y ~ x}
or \code{x1 ~~ x2}.
If \code{NULL}, then
the values of all edges will be
retrieved.}

\item{attribute_name}{The name of
the attribute from which the values
will be retrieved.}

\item{check_direction}{If \code{FALSE},
the direction of an edge is ignored.
For example, both \code{y ~ x} and \code{x ~ y}
will affect \code{y ~ x}, \code{x ~ y}, and
\code{y ~~ x}. Useful when we want to
identify an edge regardless of its
direction and whether it is directional.}
}
\value{
A named vector of the values.
}
\description{
Get the values of
an arbitrary
attribute of selected edges.
}
\details{
From a \code{\link[qgraph:qgraph]{qgraph::qgraph()}}
object generated by
\code{\link[semPlot:semPaths]{semPlot::semPaths()}}, get the
values of an attribute of selected
edges.

This function is designed to be a
general one that gets the values of the
attributes named by the user.
\subsection{Setting the value of \code{values}}{

This argument can be set in three ways.

For a named vector, the name of an
element should be the path as
specified by \link[lavaan:model.syntax]{lavaan::model.syntax}
or as appeared in
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}.

For example, if the attributes to be
changed are the colors of selected
edges, to change the color of the
path regressing \code{y} on \code{x}, the name
should be \code{"y ~ x"}. To change the
color of the covariance between \code{x1}
and \code{x2}, the name should be \code{"x1 ~~ x2"}. Therefore, \code{c("y ~ x1" = "red", "x1 ~~ x2" = "blue")} changes the
colors of the path from \code{x1} to \code{y}
and the covariance between \code{x1} and
\code{x2} to \code{"red"} and \code{"blue"},
respectively.

The order of the two nodes \emph{may}
matter for covariances. Therefore, if
the attribute of a covariance is not
changed, try switching the order of
the two nodes.

For a list of named lists, each named
list should have three named values:
\code{from}, \code{to}, and \code{new_value}. The
attribute of the edge from \code{from} to
\code{to} will be set to \code{new_value}.

The second approach is no longer
recommended, though kept for backward
compatibility.

The last approach is setting \code{values}
to a one-element vector with \emph{no} \emph{name}.
All edges in plot will then have the
selected attributes set to this value.
}
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
lavaan::parameterEstimates(fit_pa)[, c("lhs", "op", "rhs", "est", "pvalue")]
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)

get_edge_attribute(
   p_pa,
   edges = c("x2 ~~ x1", "x4 ~ x1"),
   attribute_name = "labels")

get_edge_attribute(
   p_pa,
   edges = c("x2 ~~ x1", "x4 ~ x1"),
   attribute_name = "color")


}
