% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_attribute.R
\name{get_node_attribute}
\alias{get_node_attribute}
\title{Get the Attribute Values of Selected Nodes}
\usage{
get_node_attribute(
  semPaths_plot,
  nodes = NULL,
  attribute_name = NULL,
  check_nodes = TRUE
)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions. Unlike some
other functions, this function does
\emph{not} support a list of \link[qgraph:qgraph]{qgraph::qgraph} objects.
If \code{semPaths_plot} is a list of
\link[qgraph:qgraph]{qgraph::qgraph} objects, only the
first object will be processed. Use
functions such as \code{\link[=lapply]{lapply()}} manually
to get the attribute values from a
list of \link[qgraph:qgraph]{qgraph::qgraph} objects.}

\item{nodes}{A character vector of
the names of nodes. If \code{NULL}, then
the values of all nodes will be
retrieved.}

\item{attribute_name}{The name of
the attribute from which the values
will be retrieved.}

\item{check_nodes}{Logical. If \code{TRUE}
and at least one node specified in
\code{values} are not in \code{semPaths_plot}.}
}
\value{
A named vector of the values.
}
\description{
Get the values of
an arbitrary attributes
of selected nodes.
}
\details{
From a \link[qgraph:qgraph]{qgraph::qgraph}
object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, and get the
values of an attribute of selected
nodes.

This function is designed to be a
general one that gets the values of the
attributes named by the user.
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
lavaan::parameterEstimates(fit_pa)[, c("lhs", "op", "rhs", "est", "pvalue")]
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)

get_node_attribute(
   p_pa,
   nodes = c("x1", "x3"),
   attribute_name = "color"
)
get_node_attribute(
   p_pa,
   attribute_name = "label.cex"
)

}
