% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sem_example}
\alias{sem_example}
\title{Sample dataset sem_example}
\format{
An object of class \code{data.frame} with 200 rows and 14 columns.
}
\usage{
sem_example
}
\description{
A sample dataset for fitting a latent variable model.
}
\details{
Fourteen variables (x01 to x14), 100 cases.

Sample model to fit (in lavaan notation)\preformatted{mod <-
 'f1 =~ x01 + x02 + x03
  f2 =~ x04 + x05 + x06 + x07
  f3 =~ x08 + x09 + x10
  f4 =~ x11 + x12 + x13 + x14
  f3 ~  f1 + f2
  f4 ~  f1 + f3
 '
}
}
\keyword{datasets}
