% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semptools-package.R
\docType{package}
\name{semptools-package}
\alias{semptools}
\alias{semptools-package}
\title{semptools: Customizing Structural Equation Modelling Plots}
\description{
Most function focus on specific ways to customize a graph. They use a 'qgraph' output as the first argument, and return a modified 'qgraph' object. This allows the functions to be chained by a pipe operator.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/semptools/}
  \item Report bugs at \url{https://github.com/sfcheung/semptools/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Mark Hok Chio Lai \email{marklhc@gmail.com } (\href{https://orcid.org/0000-0002-9196-7406}{ORCID})
}

}
\keyword{internal}
