% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edge_label_position.R
\name{set_edge_label_attributes}
\alias{set_edge_label_attributes}
\alias{set_edge_label_position}
\alias{set_edge_label}
\alias{set_edge_label_bg}
\title{Set Edge Label Attributes of
Selected Edges}
\usage{
set_edge_label_position(
  semPaths_plot,
  position_list = NULL,
  check_direction = TRUE
)

set_edge_label(semPaths_plot, label_list = NULL, check_direction = TRUE)

set_edge_label_bg(semPaths_plot, label_bg_list = NULL, check_direction = TRUE)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions.}

\item{position_list}{The new
position. See 'Details' on how to set
this argument. If a list of named
list is used, the element for new
position should be named
\code{new_position}.}

\item{check_direction}{If \code{FALSE},
the direction of an edge is ignored.
For example, both \code{y ~ x} and \code{x ~ y}
will affect \code{y ~ x}, \code{x ~ y}, and
\code{y ~~ x}. Useful when we want to
change an edge regardless of its
direction and whether it is directional.}

\item{label_list}{The new label. See
'Details' on how to set this
argument. If a list of named list is
used, the element for new label
should be named \code{new_label}.}

\item{label_bg_list}{The new
background color. See 'Details' on
how to set this argument. If a list
of named list is used, the element
for new background color should be
named \code{new_label_bg}.}
}
\description{
Set the attributes of
selected edges, such as the label
position.
}
\details{
Modify a \link[qgraph:qgraph]{qgraph::qgraph}
object generated by
\link[semPlot:semPaths]{semPlot::semPaths} and change the
attributes of the edge label of
selected edges.

Currently, the following attributes
are supported:
\itemize{
\item Edge label position.
\item Edge label content.
\item Edge label background color.
}
\subsection{How to specify the values}{

There are three approach to
specify the value.

First, the values can be a named
vector of the values, the recommended
approach. The name of an
element should be the path as
specified by \link[lavaan:model.syntax]{lavaan::model.syntax}
or as appeared in
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}. For
example, to change position of the
edge label of the path regressing \code{y}
on \code{x}, the name should be \code{"y ~ x"}.

Second, the values can be specified
by a list of named lists, each named
list should have three named values:
\code{from}, \code{to}, and the new value (name
depends on the attribute). The value
of the edge label of the edge from
\code{from} to \code{to} will be set to this
new value. For example, for
\code{\link[=set_edge_label_position]{set_edge_label_position()}},
\code{list(list(from = "x1", to = "y", new_position = .2), list(from = "x2", to = "y", new_position = .7))} is
equivalent to the named vector above.

The second approach is no longer
recommended. It was supported for
backward compatibility.

Third, the value can be a single
value, which will be used to set the
attribute of all edges.
}

\subsection{Setting Edge Position}{

When setting the edge position,
the value is the position. The
mid-point of the edge is 0.5. The
closer the value to 1, the closer the
label to the left-hand-side node (\code{y}
in this example). The closer the
value to 0, the close the label to
the right-hand-side node (\code{x} in this
example). For example, \code{c("y ~ x1" = .2, "y ~ x2" = .7)} moves the path
coefficient from \code{x1} to \code{y} closer
to \code{x}, and the path coefficient from
\code{x2} to \code{y} closer to \code{y}.

label positions for selected edges
changed.
}
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
lavaan::parameterEstimates(fit_pa)[, c("lhs", "op", "rhs", "est", "pvalue")]
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)

# ==== set_edge_label_position ====

my_position_vector <- c(
  "x3 ~ x2" = .25,
  "x4 ~ x1" = .75
)
p_pa2v <- set_edge_label_position(
             p_pa,
             my_position_vector
           )
plot(p_pa2v)

# This approach is no longer recommended
my_position_list <- list(
                     list(
                       from = "x2",
                       to = "x3",
                       new_position = .25
                      ),
                     list(
                       from = "x1",
                       to = "x4",
                       new_position = .75)
                     )
p_pa2l <- set_edge_label_position(
             p_pa,
             my_position_list
           )
plot(p_pa2l)


# ==== set_edge_label ====

my_label_vector <- c(
  "x3 ~ x2" = "a",
  "x4 ~ x1" = "b"
)
p_pa3v <- set_edge_label(
             p_pa,
             my_label_vector
           )
plot(p_pa3v)

# This approach is no longer recommended
my_label_list <- list(
                     list(
                       from = "x2",
                       to = "x3",
                       new_label = "a"
                      ),
                     list(
                       from = "x1",
                       to = "x4",
                       new_position = "a")
                     )
p_pa3l <- set_edge_label(
             p_pa,
             my_label_list
           )
plot(p_pa3l)


# ==== set_edge_label_bg ====

my_label_bg_vector <- c(
  "x3 ~ x2" = "red",
  "x4 ~ x1" = "#00FF00"
)
p_pa4v <- set_edge_label_bg(
             p_pa,
             my_label_bg_vector
           )
plot(p_pa4v)

# This approach is no longer recommended
my_label_bg_list <- list(
                     list(
                       from = "x2",
                       to = "x3",
                       new_label_bg = "red"
                      ),
                     list(
                       from = "x1",
                       to = "x4",
                       new_label_bg = "#00FF00")
                     )
p_pa4l <- set_edge_label_bg(
             p_pa,
             my_label_bg_list
           )
plot(p_pa4l)


}
