% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edge_width.R
\name{set_edge_line_width}
\alias{set_edge_line_width}
\title{Set the Line Widths of Selected Edges}
\usage{
set_edge_line_width(
  semPaths_plot,
  line_width_list = NULL,
  how = c("ratio", "value"),
  check_direction = TRUE
)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions.
It can also
be a list of \link[qgraph:qgraph]{qgraph::qgraph} objects,
probably though not necessarily from
a multigroup model. If it is a list
of \link[qgraph:qgraph]{qgraph::qgraph} objects, then
the function will be applied to all
the objects.}

\item{line_width_list}{A named vector or a
list of named list. See the Details
section on how to set this argument.}

\item{how}{How the width will be changed.
If \code{"ratio"}, then the new width is
the original width
multiplied by the supplied value.
If \code{"value"}, then the new width is
set to the supplied value.}

\item{check_direction}{If \code{FALSE},
the direction of an edge is ignored.
For example, both \code{y ~ x} and \code{x ~ y}
will affect \code{y ~ x}, \code{x ~ y}, and
\code{y ~~ x}. Useful when we want to
change an edge regardless of its
direction and whether it is directional.}
}
\value{
A \link[qgraph:qgraph]{qgraph::qgraph} based on
the original one, with widths for
selected edges changed.
If \code{semPaths_plot} is a list of
\link[qgraph:qgraph]{qgraph::qgraph} objects, then a
list of processed \link[qgraph:qgraph]{qgraph::qgraph} objects
will be returned.
}
\description{
Set the line widths of
selected edges.
}
\details{
Modified a \link[qgraph:qgraph]{qgraph::qgraph}
object generated by
\link[semPlot:semPaths]{semPlot::semPaths} and change the
widths of selected edges.
\subsection{Setting the value of \code{line_width_list}}{

This argument can be set in three ways.

For a named vector, the name of an
element should be the path as
specified by \link[lavaan:model.syntax]{lavaan::model.syntax}
or as appeared in
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}.

For example, to change the line width of the
path regressing \code{y} on \code{x}, the name
should be \code{"y ~ x"}. To change the
line type of the covariance between \code{x1}
and \code{x2}, the name should be \code{"x1 ~~ x2"}. Therefore, \code{c("y ~ x1" = 2, "x1 ~~ x2" = 3)} changes the
line widths of the path from \code{x1} to \code{y}
and the covariance between \code{x1} and
\code{x2}by 2 and 3,
respectively.

The order of the two nodes \emph{may}
matter for covariances. Therefore, if
the attribute of a covariance is not
changed, try switching the order of
the two nodes.

For a list of named lists, each named
list should have three named values:
\code{from}, \code{to}, and \code{new_line_width}. The
attribute of the edge from \code{from} to
\code{to} will be set to \code{new_line_width}.

The second approach is no longer
recommended, though kept for backward
compatibility.

The third approach is one single
\emph{unnamed} value. This value will be
applied to all edges.
}
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
lavaan::parameterEstimates(fit_pa)[, c("lhs", "op", "rhs", "est", "pvalue")]
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)

my_line_width_vector <- c(
   "x2 ~~ x1" = 2,
   "x4 ~ x1" = 3
)

p_pa2v <- set_edge_line_width(
 p_pa,
 my_line_width_vector)
plot(p_pa2v)

my_line_width_list <- list(
   list(
     from = "x1",
     to = "x2",
     new_line_width = 2),
   list(
     from = "x1",
     to = "x4",
     new_line_width =  3)
 )

p_pa2l <- set_edge_line_width(
   p_pa,
   my_line_width_list
 )
plot(p_pa2l)

}
