% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_graph_attributes.R
\name{set_graph_attributes}
\alias{set_graph_attributes}
\alias{set_graph_margins}
\alias{node_labels_equal_scale}
\alias{set_node_labels_equal_scale}
\title{Set Selected Attributes of a
Plot}
\usage{
set_graph_margins(
  semPaths_plot,
  bottom,
  left,
  top,
  right,
  how = c("ratio", "value")
)

node_labels_equal_scale(semPaths_plot, equal_scale)

set_node_labels_equal_scale(semPaths_plot)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions.
It can also
be a list of \link[qgraph:qgraph]{qgraph::qgraph} objects,
probably though not necessarily from
a multigroup model. If it is a list
of \link[qgraph:qgraph]{qgraph::qgraph} objects, then
the function will be applied to all
the objects.}

\item{bottom, left, right, top}{The values
for the corresponding sides. Please
refer to 'Details' on how these values
are used. Their default values depend
on \code{how}: Equal to 1 if \code{how = "ratio"},
and equal to the original values if
\code{how = "value"}.}

\item{how}{How the width will be changed.
If \code{"ratio"}, then the new width is
the original width
multiplied by the supplied value.
If \code{"value"}, then the new width is
set to the supplied value.}

\item{equal_scale}{Logical. Set the value
\code{label.scale.equal} of the plot. If
not supplied, it will not change the
existing value.}
}
\description{
Set some attributes of
a plot.
}
\details{
Functions that modify a \link[qgraph:qgraph]{qgraph::qgraph}
object generated by
\link[semPlot:semPaths]{semPlot::semPaths} and change the
widths of selected attributes.

For numerical attributes, there are
two possible modes.

If \code{how = "ratio"}, then the new value
is equal to the old value multiplied
by the user value. For example, if
the original margins are \code{c(2, 3, 4, 5)}
and the user value is 2, then the new
margins are \code{c(4, 6, 8, 10)}. If the
user values are \code{c(3, 1, 1, 1)}, then
the new margins are \code{c(6, 3, 4, 5)}.

If \code{how = "value"}, then the new value
is equal to the user value. For example, if
the original margins are \code{c(2, 3, 4, 5)}
and the user value is 2, then the new
margins are \code{c(2, 2, 2, 2)}. If
the user values are \code{c(1, 3, 5, 7)},
then the new margins are \code{c(1, 3, 5, 7)}.

The function \code{\link[=set_graph_margins]{set_graph_margins()}} is
for modifying the margins of a generated
plot. It can be useful for increasing
or decreasing one or more margins of
a plot without regenerating it.

The function \code{\link[=node_labels_equal_scale]{node_labels_equal_scale()}} is
for setting the value \code{label.scale.equal}
of a plot. By default, \code{label.scale.equal}
may be \code{FALSE}, and the labels of nodes
may be of different font sizes.
Use this function, with \code{equal_scale}
set to \code{TRUE}, will make all node labels
have the same font size.

The function \code{\link[=set_node_labels_equal_scale]{set_node_labels_equal_scale()}}
is just a wrapper of \code{\link[=node_labels_equal_scale]{node_labels_equal_scale()}}
with \code{equal_scale} set to \code{TRUE}.
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
lavaan::parameterEstimates(fit_pa)[, c("lhs", "op", "rhs", "est", "pvalue")]
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)


# ==== set_graph_margins ====

p2 <- set_graph_margins(p_pa, bottom = 2)
plot(p2)

p3 <- set_graph_margins(p_pa, bottom = 2, how = "value")
plot(p3)



# ==== node_labels_equal_scale ====

dat <- pa_example
colnames(dat)[1] <- "Mediator"
mod_pa2 <-
  'x1 ~~ x2
   Mediator ~  x1 + x2
   x4 ~  x1 + Mediator
  '
fit_pa2 <- lavaan::sem(mod_pa2, dat)
m2 <- matrix(c("x1",   NA,   NA,
                NA, "Mediator", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa2 <- semPlot::semPaths(fit_pa2, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            sizeMan = 10,
            layout = m2)

p4 <- node_labels_equal_scale(p_pa2, TRUE)
plot(p4)

p5 <- set_node_labels_equal_scale(p_pa2)
plot(p5)

}
