% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_border.R
\name{set_node_border}
\alias{set_node_border}
\alias{set_node_border_color}
\alias{set_node_border_width}
\title{Set the Attributes of Node
Borders}
\usage{
set_node_border_color(semPaths_plot, border_color = NULL, check_nodes = TRUE)

set_node_border_width(
  semPaths_plot,
  border_width = NULL,
  how = c("ratio", "value"),
  check_nodes = TRUE
)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions.
It can also
be a list of \link[qgraph:qgraph]{qgraph::qgraph} objects,
probably though not necessarily from
a multigroup model. If it is a list
of \link[qgraph:qgraph]{qgraph::qgraph} objects, then
the function will be applied to all
the objects.}

\item{border_color}{A named vector or a
list of named list. See the Details
section on how to set this argument.}

\item{check_nodes}{Logical. If \code{TRUE}
and at least one node specified in
\code{values} are not in \code{semPaths_plot}.}

\item{border_width}{A named vector or a
list of named list. See the Details
section on how to set this argument.}

\item{how}{How the width will be changed.
If \code{"ratio"}, then the new width is
the original width
multiplied by the supplied value.
If \code{"value"}, then the new width is
set to the supplied value.}
}
\value{
A \link[qgraph:qgraph]{qgraph::qgraph} based on
the original one, with the corresponding
attribute of
of selected nodes changed.
If \code{semPaths_plot} is a list of
\link[qgraph:qgraph]{qgraph::qgraph} objects, then a
list of processed \link[qgraph:qgraph]{qgraph::qgraph} objects
will be returned.
}
\description{
Set the attributes of
the borders of selected nodes.
}
\details{
Modify a \link[qgraph:qgraph]{qgraph::qgraph}
object generated by
\link[semPlot:semPaths]{semPlot::semPaths} and change the
attributes of the borders of selected
nodes.
\subsection{Setting the value of \code{border_color}}{

and \code{border_width}

This argument can be set in three
ways.

For a named vector, the name of an
element should be the nodes for which
the attributes are to be changed. The
names need to the \emph{displayed} \emph{names}
if plotted, which may be different
from the names in the model.

For example, to change the color of
\code{x}, the name should be \code{"x"}.
Therefore, \code{c("y" = "red", "x" = "blue")} changes the colors of the
nodes \code{y} and \code{x} to \code{"red"} and
\code{"blue"}, respectively.

For a list of named lists, each named
list should have two named values:
\code{node}, plus \code{new_border_color} or
\code{new_border_width}.

The second approach is no longer
recommended, though kept for backward
compatibility.

The last approach is setting
\code{border_color} or
\code{border_width} to a one-element
vector with \emph{no} \emph{name}. All nodes in
plot will then have the corresponding
attributes set to this value.
}

\subsection{Note on Border Width}{

There are also two modes for changing
border width. If \code{how = "ratio"},
then the new border width is equal to
the original border width multiplied
by the supplied value. For example,
if the supplied value is 2, and the
original border width is 1.5, the new
border width is 3.

If \code{how = "value"}, then the new
border width is set to the user
supplied value. For example, if the
supplied value is 2, then the new
border width is 2, regardless of the
original border width.
}
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)

my_border_color_vector <- c(x3 = "red", x4 = "blue")

p_pa2v <- set_node_border_color(p_pa, my_border_color_vector)
plot(p_pa2v)

p_pa2l <- set_node_border_color(p_pa, "green")
plot(p_pa2l)

my_border_width_vector <- c(x3 = 2, x4 = 3)

p_pa3v <- set_node_border_width(p_pa, my_border_width_vector)
plot(p_pa3v)

p_pa3l <- set_node_border_width(p_pa, 3)
plot(p_pa3l)

}
