% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_color.R
\name{set_node_color}
\alias{set_node_color}
\title{Set the Colors of Selected Nodes}
\usage{
set_node_color(semPaths_plot, colors = NULL)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions.}

\item{colors}{A named vector or a
list of named list. See the Details
section on how to set this argument.}
}
\value{
A \link[qgraph:qgraph]{qgraph::qgraph} based on
the original one, with the colors
of selected nodes changed.
}
\description{
Set the colors
of selected nodes.
}
\details{
Modify a \link[qgraph:qgraph]{qgraph::qgraph}
object generated by
\link[semPlot:semPaths]{semPlot::semPaths} and change the
colors of selected
nodes.
\subsection{Setting the value of \code{color_list}}{

This argument can be set in three ways.

For a named vector, the name of an
element should be the nodes for which
the colors are to be changed.
The names need to the \emph{displayed}
\emph{names} if plotted, which may be
different from the names in the model.

For example, to change the color of \code{x},
the name
should be \code{"x"}. Therefore,
\code{c("y" = "red", "x" = "red")} changes
the colors of the nodes \code{y} and \code{x}
to \code{"red"} and \code{"blue"},
respectively.

For a list of named lists, each named
list should have two named values:
\code{node} and \code{new_color}. The
color of \code{node}
will be set to \code{new_color}.

The second approach is no longer
recommended, though kept for backward
compatibility.

The last approach is setting \code{color_list}
to a one-element vector with \emph{no} \emph{name}.
All nodes in plot will then have the
colors set to this value.
}
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)

my_color_vector <- c(x3 = "red", x4 = "blue")

p_pa2v <- set_node_color(p_pa, my_color_vector)
plot(p_pa2v)

p_pa2l <- set_node_color(p_pa, "green")
plot(p_pa2l)

}
