% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_label.R
\name{set_node_label}
\alias{set_node_label}
\alias{set_node_label_size}
\alias{set_node_label_color}
\title{Set Label Attributes of Selected Nodes}
\usage{
set_node_label_size(semPaths_plot, values = NULL)

set_node_label_color(semPaths_plot, values = NULL)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions.
It can also
be a list of \link[qgraph:qgraph]{qgraph::qgraph} objects,
probably though not necessarily from
a multigroup model. If it is a list
of \link[qgraph:qgraph]{qgraph::qgraph} objects, then
the function will be applied to all
the objects.}

\item{values}{A named vector or a
list of named list. See the Details
section on how to set this argument.}
}
\value{
A \link[qgraph:qgraph]{qgraph::qgraph} based on
the original one, with the sizes
of selected nodes changed.
If \code{semPaths_plot} is a list of
\link[qgraph:qgraph]{qgraph::qgraph} objects, then a
list of processed \link[qgraph:qgraph]{qgraph::qgraph} objects
will be returned.
}
\description{
Set the attributes of labels
of selected nodes.
}
\details{
Modify a \link[qgraph:qgraph]{qgraph::qgraph}
object generated by
\link[semPlot:semPaths]{semPlot::semPaths} and change the
label attributes of selected
nodes.
\subsection{Attributes Supported}{

The function \code{\link[=set_node_label_size]{set_node_label_size()}}
set the attribute \code{label.cex} of
the plot, corresponding to the argument
\code{label.cex} of \code{\link[semPlot:semPaths]{semPlot::semPaths()}}.
It controls the size of the label of
a node.

The function \code{\link[=set_node_label_color]{set_node_label_color()}}
set the attribute \code{label.color} of
the plot, corresponding to the argument
\code{label.color} of \code{\link[semPlot:semPaths]{semPlot::semPaths()}}.
It controls the color of the label of
a node.
}

\subsection{Setting the Label-Related Attributes}{

These arguments can be set in three ways.

For a named vector, the name of an
element should be the nodes for which
the attributes are to be changed.
The names need to the \emph{displayed}
\emph{names} if plotted, which may be
different from the names in the model.

For example, to change the label-related attribute of \code{x},
the name
should be \code{"x"}. Therefore,
\code{c("y" = 2, "x" = 3)} changes
the selected attributes of the nodes \code{y} and \code{x}
to 2 and 3,
respectively.

For a list of named lists, each named
list should have two named values:
\code{node} and \code{new_value}. The
size-related attribute of \code{node}
will changed based on \code{new_value}.

The second approach is no longer
recommended, though kept for backward
compatibility.

The last approach is setting
the size-related attribute
to a one-element vector with \emph{no} \emph{name}.
All nodes in plot will then have their
sizes changed based on this value.
}
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)

p_pa2v <- set_node_label_size(p_pa, c(x1 = 5, x2 = 10))
plot(p_pa2v)

p_pa2l <- set_node_label_color(p_pa, "red")
plot(p_pa2l)

}
