% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_size.R
\name{set_node_size}
\alias{set_node_size}
\alias{set_node_width}
\alias{set_node_height}
\alias{set_node_shape}
\title{Set the Sizes of Selected Nodes}
\usage{
set_node_size(
  semPaths_plot,
  values = NULL,
  how = c("ratio", "value"),
  check_nodes = TRUE
)

set_node_width(
  semPaths_plot,
  values = NULL,
  how = c("ratio", "value"),
  check_nodes = TRUE
)

set_node_height(
  semPaths_plot,
  values = NULL,
  how = c("ratio", "value"),
  check_nodes = TRUE
)

set_node_shape(semPaths_plot, values = NULL, check_nodes = TRUE)
}
\arguments{
\item{semPaths_plot}{A
\link[qgraph:qgraph]{qgraph::qgraph} object generated by
\link[semPlot:semPaths]{semPlot::semPaths}, or a similar
qgraph object modified by other
\link{semptools} functions.}

\item{values}{A named vector or a
list of named list. See the Details
section on how to set this argument.}

\item{how}{How the width will be changed.
If \code{"ratio"}, then the new width is
the original width
multiplied by the supplied value.
If \code{"value"}, then the new width is
set to the supplied value.}

\item{check_nodes}{Logical. If \code{TRUE}
and at least one node specified in
\code{values} are not in \code{semPaths_plot}.}
}
\value{
A \link[qgraph:qgraph]{qgraph::qgraph} based on
the original one, with the sizes
of selected nodes changed.
}
\description{
Set the sizes
of selected nodes.
}
\details{
Modify a \link[qgraph:qgraph]{qgraph::qgraph}
object generated by
\link[semPlot:semPaths]{semPlot::semPaths} and change the
sizes of selected
nodes.
\subsection{Setting the Size-Related Attributes}{

These arguments can be set in three ways.

For a named vector, the name of an
element should be the nodes for which
the attributes are to be changed.
The names need to the \emph{displayed}
\emph{names} if plotted, which may be
different from the names in the model.

For example, to change the size-related attribute of \code{x},
the name
should be \code{"x"}. Therefore,
\code{c("y" = 2, "x" = 3)} changes
the selected attributes of the nodes \code{y} and \code{x}
to 2 and 3,
respectively.

For a list of named lists, each named
list should have two named values:
\code{node} and \code{new_value}. The
size-related attribute of \code{node}
will changed based on \code{new_value}.

The second approach is no longer
recommended, though kept for backward
compatibility.

The last approach is setting
the size-related attribute
to a one-element vector with \emph{no} \emph{name}.
All nodes in plot will then have their
sizes changed based on this value.
}

\subsection{Shape, Width, and Height}{

How \code{width} and \code{height} are used
depends on \code{shape}. The supported
values of \code{shape} depends on
\code{qgraph}. Common values are
are \code{"square"}, \code{"rectangle"},
\code{"circle"}, "rectangle", and \code{"ellipse"}. For
\code{"square"} and \code{"circle"}, only the
value of \code{width} is used to determine
the size of the node.
}

\subsection{Note on Changing a Value}{

There are also two modes for changing
an attribute. If \code{how = "ratio"},
then the new value, such as size,
is equal to
the original value multiplied
by the supplied value. For example,
if the supplied value is 2, and the
original size is 5, the new
size is 10.

If \code{how = "value"}, then the new
value is set to the user
supplied value. For example, if the
supplied value is 15, then the new
value is 15, regardless of the
original value.
}
}
\examples{
mod_pa <-
  'x1 ~~ x2
   x3 ~  x1 + x2
   x4 ~  x1 + x3
  '
fit_pa <- lavaan::sem(mod_pa, pa_example)
m <- matrix(c("x1",   NA,   NA,
                NA, "x3", "x4",
              "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
            style = "ram",
            nCharNodes = 0, nCharEdges = 0,
            layout = m)

p_pa2v <- set_node_size(
 p_pa, c(x1 = 5, x2 = 10),
 how = "value"
)
plot(p_pa2v)

p_pa2v2 <- set_node_size(
 p_pa, c(x1 = 0.5, x2 = 2),
 how = "ratio"
)
plot(p_pa2v2)

p_pa2l <- set_node_shape(p_pa, "circle")
plot(p_pa2l)

}
