% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setlabels.R
\name{setlabels}
\alias{setlabels}
\title{Set labels based on p-value}
\usage{
setlabels(lavaan_out, alpha = 0.05, standardized = FALSE, digits = 2)
}
\arguments{
\item{lavaan_out}{A lavaan output.}

\item{alpha}{The level of significance used to check each p-value. Default is .05.}

\item{standardized}{Whether the standardized solution will be plotted. This determined 
how the labels for dependent variable residual variances will be converted.
If TRUE, they will be converted to R-squares. If FALSE, they will be converted
to proportions of variance explained. Default is FALSE.}

\item{digits}{The number of digits used to format the labels.}
}
\value{
A lavaan output based on lavaan_out, with labels formatted based on p-values.
}
\description{
Set labels based on p-value
}
\details{
Set labels based on p-value

 This function is deprecated. Modifying the lavaan object is not a reliable 
 way to edit the qgraph. The latest workflow is to edit the qgrpah directly.
}
\examples{
\dontrun{
mod <- 
 'x1 ~~ x2
  x3 ~  x1 + x2
  x4 ~  x1 + x3
 '
fit_pa <- lavaan::sem(mod, pa_example)
out_pa <- setlabels(fit_pa)
lavaan::parameterEstimates(out_pa)[, c("lhs", "op", "rhs", 
                                       "label", "est", "pvalue")]
out2_pa <- setlabels(fit_pa, standardized = TRUE)
lavaan::parameterEstimates(out2_pa, 
standardized = TRUE)[, c("lhs", "op", "rhs", "label", "std.all", "pvalue")]

mod <- 
 'f1 =~ x01 + x02 + x03
  f2 =~ x04 + x05 + x06 + x07
  f3 =~ x08 + x09 + x10
  f4 =~ x11 + x12 + x13 + x14
 '
fit_cfa <- lavaan::sem(mod, cfa_example)
out_cfa <- setlabels(fit_cfa)
lavaan::parameterEstimates(out_cfa)[, c("lhs", "op", "rhs", 
                                       "label", "est", "pvalue")]
out2_cfa <- setlabels(fit_cfa, standardized = TRUE)
lavaan::parameterEstimates(out2_cfa, 
standardized = TRUE)[, c("lhs", "op", "rhs", "label", "std.all", "pvalue")]

mod <- 
 'f1 =~ x01 + x02 + x03
  f2 =~ x04 + x05 + x06 + x07
  f3 =~ x08 + x09 + x10
  f4 =~ x11 + x12 + x13 + x14
  f3 ~  f1 + f2
  f4 ~  f1 + f3
 '
fit_sem <- lavaan::sem(mod, sem_example)
out_sem <- setlabels(fit_sem)
lavaan::parameterEstimates(out_sem)[, c("lhs", "op", "rhs", 
                                       "label", "est", "pvalue")]
out2_sem <- setlabels(fit_sem, standardized = TRUE)
lavaan::parameterEstimates(out2_sem, 
standardized = TRUE)[, c("lhs", "op", "rhs", "label", "std.all", "pvalue")]


semPlot::semPaths(setlabels(fit_pa, alpha = .05), 
       whatLabels="name", nCharNodes = 0, nCharEdges = 0)  
       whatLabels="name", nCharNodes = 0, nCharEdges = 0)  
       whatLabels="name", nCharNodes = 0, nCharEdges = 0)  
}
}
