% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_many_controls.R
\docType{data}
\name{pa_example_3covs}
\alias{pa_example_3covs}
\title{Sample dataset pa_example_3covs}
\format{
An object of class \code{data.frame} with 100 rows and 7 columns.
}
\usage{
pa_example_3covs
}
\description{
A sample dataset for fitting a path analysis model, with three
control variables.
}
\details{
Four variables (x1 to x4), and three control variables (
cov1, cov2, cov3), 100 cases.

Sample model to fit (in \link[lavaan:model.syntax]{lavaan::model.syntax} notation)\preformatted{mod <-
'
x3 ~  x1 + x2 + cov1 +cov2 + cov3
x4 ~  x1 + x3 + cov1 +cov2 + cov3
'
}
}
\keyword{datasets}
