% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_resid.R
\name{rotate_resid}
\alias{rotate_resid}
\title{Rotate the residuals of selected nodes}
\usage{
rotate_resid(semPaths_plot, rotate_resid_list = NULL)
}
\arguments{
\item{semPaths_plot}{A qgraph object generated by semPaths, or a similar qgrpah
ojbect modified by other semptools functions.}

\item{rotate_resid_list}{A list of named lists. Each named list should have two 
named values: node and rotate. The position of the residual
of \code{node} will be placed at \code{rotate}, in degree.
The 12 o'clock position is zero degree. Positive degree 
denotes clockwise rotation, and negative degree denotes
anticlockwise rotation.}
}
\value{
A qgrpah based on the original one, with loopRotation attribtues of selected 
   nodes modified.
}
\description{
Rotate the residuals of selected nodes.
}
\details{
Modified a qgraph object generated by semPaths and rotate the residuals
         of selected edges. Currently supports "ram" and similar format only.
}
\examples{
mod_pa <- 
 'x1 ~~ x2
  x3 ~  x1 + x2
  x4 ~  x1 + x3
 '
fit_pa <- lavaan::sem(mod_pa, pa_example)
lavaan::parameterEstimates(fit_pa)[, c("lhs", "op", "rhs", "est", "pvalue")]
m <- matrix(c("x1",   NA,   NA,
               NA, "x3", "x4",
             "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
           style = "ram", 
           nCharNodes = 0, nCharEdges = 0,
           layout = m) 

my_rotate_resid_list <- list(list(node = "x3", rotate =  45),
                         list(node = "x4", rotate = -45))
 
p_pa2 <- rotate_resid(p_pa, my_rotate_resid_list)
plot(p_pa2)

}
