% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semptools-package.R
\docType{package}
\name{semptools-package}
\alias{semptools}
\alias{semptools-package}
\title{semptools: Helper Functions for Customizing Plots Generated by semPlot::semPaths}
\description{
Helper functions for customizing plots generated by semPlot::semPaths.
             Each function should focus on one specific purpose. Most functions
             should use a qgraph output as the first argument, and return
             a modified qgraph object. This allows the functions to be chained
             by the pipe operator.
}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Mark Hok Chio Lai \email{marklhc@gmail.com } (\href{https://orcid.org/0000-0002-9196-7406}{ORCID})
}

}
\keyword{internal}
