% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_curve.R
\name{set_curve}
\alias{set_curve}
\title{Set the Curve Attributes of Selected edges}
\usage{
set_curve(semPaths_plot, curve_list = NULL)
}
\arguments{
\item{semPaths_plot}{A qgraph object generated by semPaths, or a similar qgrpah
object modified by other semptools functions.}

\item{curve_list}{A list of named lists. Each named list should have three
named values: from, to, and new_curve. The curve attribute
of the edge from \code{from} to \code{to} will be set to
\code{new_curve}.  Alternatively, it can be a named
vector, with \code{\link{new_curve}} as the value, and the name
of each value being the name of the curve/path as in the
\code{\link{lavaan}} model. For example, if the curve to be modified
is the path from \code{\link{x1}} to \code{\link{x4}}, then the name is \code{x4 ~ x1},
as appeared on the parameter table of the fit.}
}
\value{
A qgrpah based on the original one, with curve attributes for selected edges
changed.
}
\description{
Set the curve attributes of selected edges.
}
\details{
Modified a qgraph object generated by semPaths and change the
curve attributes of selected edges.
}
\examples{
mod_pa <- 
 'x1 ~~ x2
  x3 ~  x1 + x2
  x4 ~  x1 + x3
 '
fit_pa <- lavaan::sem(mod_pa, pa_example)
lavaan::parameterEstimates(fit_pa)[, c("lhs", "op", "rhs", "est", "pvalue")]
m <- matrix(c("x1",   NA,   NA,
               NA, "x3", "x4",
             "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
           style = "ram", 
           nCharNodes = 0, nCharEdges = 0,
           layout = m) 

my_curve_list <- list(list(from = "x1", to = "x2", new_curve = -1),
                    list(from = "x1", to = "x4", new_curve =  1))
 
p_pa2 <- set_curve(p_pa, my_curve_list)
plot(p_pa2)

}
