% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edge_label_position.R
\name{set_edge_label_position}
\alias{set_edge_label_position}
\title{Set the positions of edge labels of selected edges}
\usage{
set_edge_label_position(semPaths_plot, position_list = NULL)
}
\arguments{
\item{semPaths_plot}{A qgraph object generated by semPaths, or a similar
qgrpah
object modified by other semptools functions.}

\item{position_list}{A list of named lists. Each named list should have three
named values: from, to, and new_position. The edge label
position
of the edge from \code{from} to \code{to} will be set to
\code{new_position}. Alternatively, it can be a named
vector, with \code{new_position} as the value, and the name
of each value being the name of the curve/path as in the
\code{lavaan} model. For example, if the curve to be modified
is the path from \code{x1} to \code{x4}, then the name is
\code{x4 ~ x1},
as appeared on the parameter table of the fit.}
}
\value{
A qgrpah based on the original one, with edge label positions for selected
edges changed.
}
\description{
Set the positions of edge labels of selected edges.
}
\details{
Modified a qgraph object generated by semPaths and change the
edge label positions of selected edges.
}
\examples{
mod_pa <-
 'x1 ~~ x2
  x3 ~  x1 + x2
  x4 ~  x1 + x3
 '
fit_pa <- lavaan::sem(mod_pa, pa_example)
lavaan::parameterEstimates(fit_pa)[, c("lhs", "op", "rhs", "est", "pvalue")]
m <- matrix(c("x1",   NA,   NA,
               NA, "x3", "x4",
             "x2",   NA,   NA), byrow = TRUE, 3, 3)
p_pa <- semPlot::semPaths(fit_pa, whatLabels="est",
           style = "ram",
           nCharNodes = 0, nCharEdges = 0,
           layout = m)

my_position_list <- list(list(from = "x2", to = "x3", new_position =  .25),
                       list(from = "x1", to = "x4", new_position =  .75))


p_pa2 <- set_edge_label_position(p_pa, my_position_list)
plot(p_pa2)

}
